+++ packages/moonstone/VideoPlayer/MediaControls.js
@@ -572,6 +572,14 @@ const MediaControlsDecorator = hoc((config, Wrapped) => {
 			paused: PropTypes.bool,
 
 			/**
+			 * A DOM Node of the player.
+			 *
+			 * @type {Node}
+			 * @public
+			 */
+			playerNode: PropTypes.node,
+
+			/**
 			 * Sets the `disabled` state on the media playback-rate control buttons; the inner pair.
 			 *
 			 * @type {Boolean}
@@ -637,8 +645,8 @@ const MediaControlsDecorator = hoc((config, Wrapped) => {
 				countReactChildren(this.props.rightComponents),
 				countReactChildren(this.props.children)
 			);
-			on('keydown', this.handleKeyDown);
-			on('keyup', this.handleKeyUp);
+			on('keydown', this.handleKeyDown, this.props.playerNode);
+			on('keyup', this.handleKeyUp, this.props.playerNode);
 		}
 
 		componentWillReceiveProps (nextProps) {
@@ -835,6 +843,7 @@ const MediaControlsDecorator = hoc((config, Wrapped) => {
 			delete props.onPlay;
 			delete props.onRewind;
 			delete props.onToggleMore;
+			delete props.playerNode;
 			delete props.setApiProvider;
 
 			return (